phone_backend/requirements.txt
fastapi
uvicorn
twilio
openai
python-dotenv
